refactor: separate some generation HTML code
