+++ client/src/helpers/generateHTML/message.js
@@ -2,117 +2,18 @@
 
 import React from 'react';
 import { Layout, Input, Button, List, Avatar, Icon, Row, Col, Badge, Popover, message, Spin, Tabs } from 'antd';
-import { messageConfig, block } from '../config/message';
-import avatarConfig from '../config/avatar';
-import { getUserAvatarUrl, getEmoji } from './../helpers/common';
-import configEmoji from '../config/emoji';
-import InfiniteScroll from 'react-infinite-scroller';
+import { messageConfig } from '../../config/message';
 import { Link } from 'react-router-dom';
-import handlersMessage from './handlersMessage';
-import { room } from '../config/room';
-import ModalSetNicknames from '../components/modals/room/ModalSetNicknames';
-
-const { TabPane } = Tabs;
+import { getUserAvatarUrl, getEmoji } from './../common';
+import { generateListTo } from './to';
+import { generateListEmoji, generateEmojiButton } from './emoji';
+import handlersMessage from './../handlersMessage';
+import configEmoji from '../../config/emoji';
+import avatarConfig from '../../config/avatar';
+import { generateReactionMsg, generateReactionUserList } from './reaction';
 
 export function getReplyMessageContent(component, message) {
-  const { nicknames, messageIdEditing } = component.state;
-  const currentUserInfo = component.props.userContext.info;
-
-  let replyMessageContent = component.createMarkupMessage(message);
-  let notificationClass = message.is_notification ? 'pre-notification' : '';
-  let isToMe =
-    replyMessageContent.__html.includes(`data-tag="[To:${currentUserInfo._id}]"`) ||
-    replyMessageContent.__html.includes(`data-tag="[rp mid=${currentUserInfo._id}]"`) ||
-    replyMessageContent.__html.includes(messageConfig.SIGN_TO_ALL);
-  let reactionOfMsg = component.reactionDupplicateCounter(message.reactions);
-
   return generateMessageHTML(component, message, true);
-}
-
-export function generateReactionMsg(component, msgId) {
-  const listReaction = configEmoji.REACTION;
-  const { t } = component.props;
-  const content = (
-    <div id="_reactionList" className="reactionSelectorTooltip">
-      <ul className="reactionSelectorTooltip__emoticonList">
-        {Object.entries(listReaction).map(([key, reaction]) => {
-          return (
-            <li className="reactionSelectorTooltip__itemContainer" key={key}>
-              <span className="reactionSelectorTooltip__item">
-                <span className="reactionSelectorTooltip__emoticonContainer">
-                  <Avatar
-                    className="reactionSelectorTooltip__emoticon image-emoji"
-                    src={getEmoji(reaction.image)}
-                    alt={key}
-                    title={t(reaction.tooltip)}
-                    onClick={() => component.handleReaction(msgId, key)}
-                  />
-                </span>
-              </span>
-            </li>
-          );
-        })}
-      </ul>
-    </div>
-  );
-
-  return content;
-}
-
-export function generateReactionUserList(component, msgId, reactionOfMsg) {
-  return (
-    <div className="reactionUserListTollTip">
-      <Tabs activeKey={component.state.activeKeyTab} onChange={component.onChangeTab}>
-        {reactionOfMsg.map((value, index) => {
-          return (
-            <TabPane
-              tab={
-                <div onClick={() => component.fetchReactionUserList(msgId, value.reaction.reaction_tag)}>
-                  <img
-                    src={getEmoji(configEmoji.REACTION[value.reaction.reaction_tag].image)}
-                    alt={value.reaction.reaction_tag}
-                    className="reactionButton__emoticon"
-                  />
-                  <span className="reactionButton__count _reactionCount">{value.count}</span>
-                </div>
-              }
-              key={index}
-            >
-              {component.contentReactionUserList(msgId, value.reaction.reaction_tag)}
-            </TabPane>
-          );
-        })}
-      </Tabs>
-    </div>
-  );
-}
-
-export function generateListEMoji(component) {
-  const listEmoji = configEmoji.EMOJI;
-  const { t } = component.props;
-  const content = (
-    <div className="member-infinite-container" style={{ width: '210px' }}>
-      <InfiniteScroll initialLoad={false} pageStart={0} loadMore={component.handleInfiniteOnLoad} useWindow={false}>
-        <div className="box-emoji">
-          {Object.entries(listEmoji).map(([key, emoji]) => {
-            return (
-              <p className="line-emoji" key={key}>
-                <Avatar
-                  className="image-emoji"
-                  src={getEmoji(emoji.image)}
-                  alt={key}
-                  title={t(emoji.tooltip)}
-                  onClick={component.handleEmoji}
-                />
-              </p>
-            );
-          })}
-        </div>
-      </InfiniteScroll>
-    </div>
-  );
-
-  return content;
 }
 
 export function generateMsgContent(component, infoUserTip) {
@@ -174,65 +75,6 @@ export function generateMsgContent(component, infoUserTip) {
   );
 }
 
-export function generateRedLine(component) {
-  const { t } = component.props;
-
-  return (
-    <div className={'timeLine__unreadLine'} ref={element => (component.attr.unreadMsgLineRef = element)}>
-      <div className="timeLine__unreadLineBorder">
-        <div className="timeLine__unreadLineContainer">
-          <div className="timeLine__unreadLineBody">
-            <span className="timeLine__unreadLineText">{t('unread_title')}</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export function generateListTo(component) {
-  const { t, allMembers, roomInfo } = component.props;
-  const currentUserInfo = component.props.userContext.info;
-  const content =
-    allMembers == [] ? (
-      <span>{t('not_data')}</span>
-    ) : (
-      <React.Fragment>
-        <div className="member-infinite-container">
-          {roomInfo.type == room.ROOM_TYPE.GROUP_CHAT && (
-            <a className="form-control to-all" href="javascript:;" onClick={handlersMessage.actionFunc.toAll}>
-              <span>{t('to_all')}</span>
-            </a>
-          )}
-          <InfiniteScroll initialLoad={false} pageStart={0} loadMore={component.handleInfiniteOnLoad} useWindow={false}>
-            <List
-              dataSource={allMembers}
-              renderItem={member => {
-                return member._id != currentUserInfo._id ? (
-                  <List.Item key={member._id}>
-                    <List.Item.Meta
-                      avatar={<Avatar size={avatarConfig.AVATAR.SIZE.SMALL} src={getUserAvatarUrl(member.avatar)} />}
-                      title={
-                        <a onClick={handlersMessage.actionFunc.toMember} href="javascript:;" data-mid={member._id}>
-                          {member.nickname ? member.nickname.nickname : member.name}
-                        </a>
-                      }
-                    />
-                  </List.Item>
-                ) : (
-                  <span />
-                );
-              }}
-            />
-          </InfiniteScroll>
-        </div>
-        <ModalSetNicknames hidePopoverTo={component.hidePopoverTo} members={allMembers} />
-      </React.Fragment>
-    );
-
-  return content;
-}
-
 export function generateMessageHTML(component, message, isGetContentOfReplyMsg = false) {
   let {
     messages,
@@ -249,7 +91,7 @@ export function generateMessageHTML(component, message, isGetContentOfReplyMsg =
   const { t, roomInfo, isReadOnly, roomId, allMembers } = component.props;
   const currentUserInfo = component.props.userContext.info;
   const showListMember = generateListTo(component);
-  const showListEmoji = generateListEMoji(component);
+  const showListEmoji = generateListEmoji(component);
   const redLine = generateRedLine(component);
   const listMember = allMembers.filter(item => item._id != currentUserInfo._id);
   let nextMsgId = null;
@@ -315,7 +157,7 @@ export function generateMessageHTML(component, message, isGetContentOfReplyMsg =
               <div className="_reaction timelineMessage__reactionDisplayContainer" style={{ display: 'flex' }}>
                 {reactionOfMsg.map(value => {
                   return configEmoji.REACTION[value.reaction.reaction_tag]
-                    ? generateEmojiHTML(component, message, value, isGetContentOfReplyMsg)
+                    ? generateEmojiButton(component, message, value, isGetContentOfReplyMsg)
                     : '';
                 })}
                 {!isGetContentOfReplyMsg && (
@@ -402,33 +244,18 @@ export function generateMessageHTML(component, message, isGetContentOfReplyMsg =
   );
 }
 
-export function generateEmojiHTML(component, message, value, isGetContentOfReplyMsg = false) {
-  return !isGetContentOfReplyMsg ? (
-    <span
-      className="reactionButton reactionButton--myReaction _sendReaction _showDescription"
-      aria-label="Remove component reaction"
-      data-reactiontype="yes"
-      onClick={() => component.handleReaction(message._id, value.reaction.reaction_tag)}
-    >
-      <img
-        src={getEmoji(configEmoji.REACTION[value.reaction.reaction_tag].image)}
-        alt={value.reaction.reaction_tag}
-        className="reactionButton__emoticon"
-      />
-      <span className="reactionButton__count _reactionCount">{value.count}</span>
-    </span>
-  ) : (
-    <span
-      className="reactionButton reactionButton--myReaction _sendReaction _showDescription"
-      aria-label="Remove component reaction"
-      data-reactiontype="yes"
-    >
-      <img
-        src={getEmoji(configEmoji.REACTION[value.reaction.reaction_tag].image)}
-        alt={value.reaction.reaction_tag}
-        className="reactionButton__emoticon"
-      />
-      <span className="reactionButton__count _reactionCount">{value.count}</span>
-    </span>
+export function generateRedLine(component) {
+  const { t } = component.props;
+
+  return (
+    <div className={'timeLine__unreadLine'} ref={element => (component.attr.unreadMsgLineRef = element)}>
+      <div className="timeLine__unreadLineBorder">
+        <div className="timeLine__unreadLineContainer">
+          <div className="timeLine__unreadLineBody">
+            <span className="timeLine__unreadLineText">{t('unread_title')}</span>
+          </div>
+        </div>
+      </div>
+    </div>
   );
 }
